requirements.txt
cffi==1.14.4
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.1.0
pycparser==2.20
pyloudnorm==0.1.0
pyparsing==2.4.7
PySoundFile==0.9.0.post1
python-dateutil==2.8.1
scipy==1.6.0
six==1.15.0
